# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OneTune music streaming API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: search
    description: Music search
  - name: tracks
    description: Track operations
  - name: playlists
    description: Playlist management
  - name: library
    description: User library
  - name: history
    description: Playback history
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /search:
    get:
      operationId: searchMusic
      tags: [search]
      summary: Search music via iTunes API
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [track, artist, album, all]
            default: all
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

  /tracks/trending:
    get:
      operationId: getTrendingTracks
      tags: [tracks]
      summary: Get trending tracks
      parameters:
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Trending tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"

  /tracks/featured:
    get:
      operationId: getFeaturedTracks
      tags: [tracks]
      summary: Get featured spotlight tracks for hero section
      responses:
        "200":
          description: Featured tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"

  /tracks/genres:
    get:
      operationId: getGenres
      tags: [tracks]
      summary: Get available music genres
      responses:
        "200":
          description: Genres list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Genre"

  /tracks/{id}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get track by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Track details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          description: Track not found

  /playlists:
    get:
      operationId: getPlaylists
      tags: [playlists]
      summary: List all user playlists
      responses:
        "200":
          description: Playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a new playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Created playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"

  /playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get playlist with its tracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist with tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistWithTracks"
        "404":
          description: Not found
    patch:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Update playlist metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistUpdate"
      responses:
        "200":
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
        "404":
          description: Not found
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /playlists/{id}/tracks:
    post:
      operationId: addTrackToPlaylist
      tags: [playlists]
      summary: Add a track to a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistTrackInput"
      responses:
        "201":
          description: Track added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistTrack"

  /playlists/{id}/tracks/{trackId}:
    delete:
      operationId: removeTrackFromPlaylist
      tags: [playlists]
      summary: Remove a track from a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  /library:
    get:
      operationId: getLibrary
      tags: [library]
      summary: Get user's saved/liked tracks
      responses:
        "200":
          description: Library tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryTrack"
    post:
      operationId: saveToLibrary
      tags: [library]
      summary: Save a track to library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryTrackInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryTrack"

  /library/{trackId}:
    delete:
      operationId: removeFromLibrary
      tags: [library]
      summary: Remove track from library
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get recently played tracks
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recently played
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"
    post:
      operationId: addToHistory
      tags: [history]
      summary: Record a played track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistoryInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"

  /stats:
    get:
      operationId: getStats
      tags: [history]
      summary: Get user listening stats (total plays, top genres, top artists)
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListeningStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Track:
      type: object
      required: [id, title, artist, artworkUrl]
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: ["string", "null"]
        artworkUrl:
          type: string
        previewUrl:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        genre:
          type: ["string", "null"]
        releaseDate:
          type: ["string", "null"]
        trackNumber:
          type: ["integer", "null"]

    Genre:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: ["string", "null"]

    SearchResults:
      type: object
      required: [tracks, artists, albums]
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        artists:
          type: array
          items:
            $ref: "#/components/schemas/Artist"
        albums:
          type: array
          items:
            $ref: "#/components/schemas/Album"

    Artist:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        artworkUrl:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]

    Album:
      type: object
      required: [id, title, artist]
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        artworkUrl:
          type: ["string", "null"]
        releaseDate:
          type: ["string", "null"]
        trackCount:
          type: ["integer", "null"]

    Playlist:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        trackCount:
          type: integer
        createdAt:
          type: string

    PlaylistWithTracks:
      type: object
      required: [id, name, createdAt, tracks]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistTrack"

    PlaylistInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        artworkUrl:
          type: string

    PlaylistUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        artworkUrl:
          type: string

    PlaylistTrack:
      type: object
      required: [id, playlistId, trackId, title, artist, artworkUrl, addedAt]
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: ["string", "null"]
        artworkUrl:
          type: string
        previewUrl:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        addedAt:
          type: string

    PlaylistTrackInput:
      type: object
      required: [trackId, title, artist, artworkUrl]
      properties:
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        artworkUrl:
          type: string
        previewUrl:
          type: string
        durationMs:
          type: integer

    LibraryTrack:
      type: object
      required: [id, trackId, title, artist, artworkUrl, savedAt]
      properties:
        id:
          type: integer
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: ["string", "null"]
        artworkUrl:
          type: string
        previewUrl:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        genre:
          type: ["string", "null"]
        savedAt:
          type: string

    LibraryTrackInput:
      type: object
      required: [trackId, title, artist, artworkUrl]
      properties:
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        artworkUrl:
          type: string
        previewUrl:
          type: string
        durationMs:
          type: integer
        genre:
          type: string

    HistoryEntry:
      type: object
      required: [id, trackId, title, artist, artworkUrl, playedAt]
      properties:
        id:
          type: integer
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        artworkUrl:
          type: string
        previewUrl:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        playedAt:
          type: string

    HistoryInput:
      type: object
      required: [trackId, title, artist, artworkUrl]
      properties:
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        artworkUrl:
          type: string
        previewUrl:
          type: string
        durationMs:
          type: integer

    ListeningStats:
      type: object
      required: [totalPlays, topGenres, topArtists]
      properties:
        totalPlays:
          type: integer
        topGenres:
          type: array
          items:
            $ref: "#/components/schemas/GenreStat"
        topArtists:
          type: array
          items:
            $ref: "#/components/schemas/ArtistStat"

    GenreStat:
      type: object
      required: [genre, count]
      properties:
        genre:
          type: string
        count:
          type: integer

    ArtistStat:
      type: object
      required: [artist, count]
      properties:
        artist:
          type: string
        count:
          type: integer
        artworkUrl:
          type: ["string", "null"]