# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneTune music streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: string;
  title: string;
  artist: string;
  /** @nullable */
  artworkUrl?: string | null;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  trackCount?: number | null;
}