# lib/api-zod/src/generated/types/artistStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneTune music streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface ArtistStat {
  artist: string;
  count: number;
  /** @nullable */
  artworkUrl?: string | null;
}