# lib/api-zod/src/generated/types/genre.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneTune music streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface Genre {
  id: string;
  name: string;
  /** @nullable */
  color?: string | null;
}