# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneTune music streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  trackId: string;
  title: string;
  artist: string;
  artworkUrl: string;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  durationMs?: number | null;
  playedAt: string;
}