# lib/api-zod/src/generated/types/libraryTrackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneTune music streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryTrackInput {
  trackId: string;
  title: string;
  artist: string;
  album?: string;
  artworkUrl: string;
  previewUrl?: string;
  durationMs?: number;
  genre?: string;
}