# lib/api-zod/src/generated/types/listeningStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneTune music streaming API
 * OpenAPI spec version: 0.1.0
 */
import type { ArtistStat } from './artistStat';
import type { GenreStat } from './genreStat';

export interface ListeningStats {
  totalPlays: number;
  topGenres: GenreStat[];
  topArtists: ArtistStat[];
}