# lib/api-zod/src/generated/types/playlistInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneTune music streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  artworkUrl?: string;
}